requirements.txt
torch
chess
pyyaml
tqdm
numpy
